Support configuration of number precision
'Integer' checkbox was ignored in the runner anyway, so it seems
safe to replace wholesale with a 'Precision' configuration option
allowing the user to specify the number of decimal places to allow.

Defaults to 0, so by default numbers are integers.

Expose the number of precision, if present, as the 'step' attribute
on the corresponding number input in the runner.

# Conflicts:
#	designer/dist/designer.js